+++ recipes/upgrade.rb
@@ -1,0 +1,11 @@
+#
+# Cookbook:: pdsoe
+# Recipe:: upgrade
+#
+# Copyright:: 2022, Mike Butler, All Rights Reserved.
+
+log 'Coming Soon' do
+  level :info
+end
+
+untag('oe_upgrade')